+++ app/src/main/java/woowacourse/payments/ui/cardcreate/PaymentCard.kt
@@ -1,0 +1,40 @@
+package woowacourse.payments.ui.cardcreate
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.unit.dp
+import woowacourse.payments.R
+
+@Composable
+fun PaymentCard(modifier: Modifier = Modifier) {
+    Box(
+        contentAlignment = Alignment.CenterStart,
+        modifier =
+            modifier
+                .shadow(8.dp)
+                .size(width = 208.dp, height = 124.dp)
+                .background(
+                    color = colorResource(id = R.color.gray_33),
+                    shape = RoundedCornerShape(5.dp),
+                ),
+    ) {
+        Box(
+            modifier =
+                Modifier
+                    .padding(start = 14.dp, bottom = 10.dp)
+                    .size(width = 40.dp, height = 26.dp)
+                    .background(
+                        color = colorResource(id = R.color.yellow_CB),
+                        shape = RoundedCornerShape(4.dp),
+                    ),
+        )
+    }
+}
